+++ src/backend/vector/scalar_mul/pippenger.rs
@@ -10,6 +10,7 @@
 #![allow(non_snake_case)]
 
 use core::borrow::Borrow;
+use core::cmp::Ordering;
 
 use crate::backend::vector::{CachedPoint, ExtendedPoint};
 use crate::edwards::EdwardsPoint;
@@ -51,7 +52,7 @@ impl VartimeMultiscalarMul for Pippenger {
 
         // Collect optimized scalars and points in a buffer for repeated access
         // (scanning the whole collection per each digit position).
-        let scalars = scalars.into_iter().map(|s| s.borrow().to_radix_2w(w));
+        let scalars = scalars.into_iter().map(|s| s.borrow().as_radix_2w(w));
 
         let points = points
             .into_iter()
@@ -70,8 +71,8 @@ impl VartimeMultiscalarMul for Pippenger {
 
         let mut columns = (0..digits_count).rev().map(|digit_index| {
             // Clear the buckets when processing another digit.
-            for i in 0..buckets_count {
-                buckets[i] = ExtendedPoint::identity();
+            for bucket in &mut buckets {
+                *bucket = ExtendedPoint::identity();
             }
 
             // Iterate over pairs of (point, scalar)
@@ -81,12 +82,16 @@ impl VartimeMultiscalarMul for Pippenger {
             for (digits, pt) in scalars_points.iter() {
                 // Widen digit so that we don't run into edge cases when w=8.
                 let digit = digits[digit_index] as i16;
-                if digit > 0 {
-                    let b = (digit - 1) as usize;
-                    buckets[b] = &buckets[b] + pt;
-                } else if digit < 0 {
-                    let b = (-digit - 1) as usize;
-                    buckets[b] = &buckets[b] - pt;
+                match digit.cmp(&0) {
+                    Ordering::Greater => {
+                        let b = (digit - 1) as usize;
+                        buckets[b] = &buckets[b] + pt;
+                    }
+                    Ordering::Less => {
+                        let b = (-digit - 1) as usize;
+                        buckets[b] = &buckets[b] - pt;
+                    }
+                    Ordering::Equal => {}
                 }
             }