+++ src/backend/vector/scalar_mul/straus.rs
@@ -12,6 +12,7 @@
 #![allow(non_snake_case)]
 
 use core::borrow::Borrow;
+use core::cmp::Ordering;
 
 use zeroize::Zeroizing;
 
@@ -53,7 +54,7 @@ impl MultiscalarMul for Straus {
 
         let scalar_digits_vec: Vec<_> = scalars
             .into_iter()
-            .map(|s| s.borrow().to_radix_16())
+            .map(|s| s.borrow().as_radix_16())
             .collect();
         // Pass ownership to a `Zeroizing` wrapper
         let scalar_digits = Zeroizing::new(scalar_digits_vec);
@@ -95,10 +96,14 @@ impl VartimeMultiscalarMul for Straus {
             Q = Q.double();
 
             for (naf, lookup_table) in nafs.iter().zip(lookup_tables.iter()) {
-                if naf[i] > 0 {
-                    Q = &Q + &lookup_table.select(naf[i] as usize);
-                } else if naf[i] < 0 {
-                    Q = &Q - &lookup_table.select(-naf[i] as usize);
+                match naf[i].cmp(&0) {
+                    Ordering::Greater => {
+                        Q = &Q + &lookup_table.select(naf[i] as usize);
+                    }
+                    Ordering::Less => {
+                        Q = &Q - &lookup_table.select(-naf[i] as usize);
+                    }
+                    Ordering::Equal => {}
                 }
             }
         }