+++ src/backend/vector/scalar_mul/vartime_double_base.rs
@@ -11,6 +11,8 @@
 
 #![allow(non_snake_case)]
 
+use core::cmp::Ordering;
+
 use crate::backend::vector::BASEPOINT_ODD_LOOKUP_TABLE;
 use crate::backend::vector::{CachedPoint, ExtendedPoint};
 use crate::edwards::EdwardsPoint;
@@ -40,16 +42,24 @@ pub fn mul(a: &Scalar, A: &EdwardsPoint, b: &Scalar) -> EdwardsPoint {
     loop {
         Q = Q.double();
 
-        if a_naf[i] > 0 {
-            Q = &Q + &table_A.select(a_naf[i] as usize);
-        } else if a_naf[i] < 0 {
-            Q = &Q - &table_A.select(-a_naf[i] as usize);
+        match a_naf[i].cmp(&0) {
+            Ordering::Greater => {
+                Q = &Q + &table_A.select(a_naf[i] as usize);
+            }
+            Ordering::Less => {
+                Q = &Q - &table_A.select(-a_naf[i] as usize);
+            }
+            Ordering::Equal => {}
         }
 
-        if b_naf[i] > 0 {
-            Q = &Q + &table_B.select(b_naf[i] as usize);
-        } else if b_naf[i] < 0 {
-            Q = &Q - &table_B.select(-b_naf[i] as usize);
+        match b_naf[i].cmp(&0) {
+            Ordering::Greater => {
+                Q = &Q + &table_B.select(b_naf[i] as usize);
+            }
+            Ordering::Less => {
+                Q = &Q - &table_B.select(-b_naf[i] as usize);
+            }
+            Ordering::Equal => {}
         }
 
         if i == 0 {